requirements.txt
streamlit
chromadb
google-generativeai
psycopg2